// src/lib/cardcom/generated/models/acquire.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * CardCom API
 * OpenAPI spec version: 11.0
 */

export type Acquire = typeof Acquire[keyof typeof Acquire];


export const Acquire = {
  Unknown: 'Unknown',
  Isracard: 'Isracard',
  CAL: 'CAL',
  Diners: 'Diners',
  AmericanExpress: 'AmericanExpress',
  Laumicard: 'Laumicard',
  CardCom: 'CardCom',
  PayPal: 'PayPal',
  Upay: 'Upay',
  PayMe: 'PayMe',
} as const;