// src/lib/cardcom/generated/models/brand.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * CardCom API
 * OpenAPI spec version: 11.0
 */

export type Brand = typeof Brand[keyof typeof Brand];


export const Brand = {
  PrivateCard: 'PrivateCard',
  MasterCard: 'MasterCard',
  Visa: 'Visa',
  Maestro: 'Maestro',
  AmericanExpress: 'AmericanExpress',
  Isracard: 'Isracard',
  JBC: 'JBC',
  Discover: 'Discover',
  Diners: 'Diners',
} as const;