// src/lib/cardcom/generated/models/cardInfo.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * CardCom API
 * OpenAPI spec version: 11.0
 */

export type CardInfo = typeof CardInfo[keyof typeof CardInfo];


export const CardInfo = {
  Israeli: 'Israeli',
  NonIsraeli: 'NonIsraeli',
  FuelCard: 'FuelCard',
  ImmediateChargeCard: 'ImmediateChargeCard',
  GiftCard: 'GiftCard',
} as const;