// src/lib/cardcom/generated/models/cardNumberEntryMode.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * CardCom API
 * OpenAPI spec version: 11.0
 */

export type CardNumberEntryMode = typeof CardNumberEntryMode[keyof typeof CardNumberEntryMode];


export const CardNumberEntryMode = {
  MagneticStip: 'MagneticStip',
  SelfService: 'SelfService',
  GasStationSelfService: 'GasStationSelfService',
  Contactless: 'Contactless',
  EmvContactless: 'EmvContactless',
  MobileContactless: 'MobileContactless',
  EmvMobileContactless: 'EmvMobileContactless',
  MobileNumber: 'MobileNumber',
  Emv: 'Emv',
  Phone: 'Phone',
  SignatureOnly: 'SignatureOnly',
  Internet: 'Internet',
  Fallback: 'Fallback',
  EmptyCandidateList: 'EmptyCandidateList',
} as const;