// src/lib/cardcom/generated/models/dealType.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * CardCom API
 * OpenAPI spec version: 11.0
 */

export type DealType = typeof DealType[keyof typeof DealType];


export const DealType = {
  Information: 'Information',
  Debit: 'Debit',
  Discharge: 'Discharge',
  ForcedCharge: 'ForcedCharge',
  CashBack: 'CashBack',
  CashTransaction: 'CashTransaction',
  Recurring: 'Recurring',
  BalanceQuery: 'BalanceQuery',
  Cancel: 'Cancel',
  Refund: 'Refund',
  Recharge: 'Recharge',
} as const;