// src/lib/cardcom/generated/models/documentType.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * CardCom API
 * OpenAPI spec version: 11.0
 */

export type DocumentType = typeof DocumentType[keyof typeof DocumentType];


export const DocumentType = {
  Error: 'Error',
  TaxInvoiceAndReceipt: 'TaxInvoiceAndReceipt',
  TaxInvoiceAndReceiptRefund: 'TaxInvoiceAndReceiptRefund',
  Receipt: 'Receipt',
  ReceiptRefund: 'ReceiptRefund',
  Quote: 'Quote',
  Order: 'Order',
  SiteCustomerOrder: 'SiteCustomerOrder',
  SiteCustomerOrderRefund: 'SiteCustomerOrderRefund',
  DeliveryNote: 'DeliveryNote',
  DeliveryNoteRefund: 'DeliveryNoteRefund',
  ProformaInvoice: 'ProformaInvoice',
  DemandForPayment: 'DemandForPayment',
  DemandForPaymentRefund: 'DemandForPaymentRefund',
  TaxInvoice: 'TaxInvoice',
  TaxInvoiceRefund: 'TaxInvoiceRefund',
  ReceiptForTaxInvoice: 'ReceiptForTaxInvoice',
  DonationReceipt: 'DonationReceipt',
  DonationReceiptRefund: 'DonationReceiptRefund',
  ReceiptForTaxInvoiceRefund: 'ReceiptForTaxInvoiceRefund',
} as const;