// src/lib/cardcom/generated/models/issuer.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * CardCom API
 * OpenAPI spec version: 11.0
 */

export type Issuer = typeof Issuer[keyof typeof Issuer];


export const Issuer = {
  NonIsrael: 'NonIsrael',
  Isracard: 'Isracard',
  CAL: 'CAL',
  Diners: 'Diners',
  AmericanExpress: 'AmericanExpress',
  JCB: 'JCB',
  Laumicard: 'Laumicard',
} as const;