// src/lib/cardcom/generated/models/paymentType.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * CardCom API
 * OpenAPI spec version: 11.0
 */

export type PaymentType = typeof PaymentType[keyof typeof PaymentType];


export const PaymentType = {
  Unknown: 'Unknown',
  Standard: 'Standard',
  SpecialCredits: 'SpecialCredits',
  ImmediateCharge: 'ImmediateCharge',
  CreditClub: 'CreditClub',
  SuperCredit: 'SuperCredit',
  InstallmentCredit: 'InstallmentCredit',
  Payments: 'Payments',
  ClubPatments: 'ClubPatments',
} as const;